Сделал компонент-контейнер для формы отправки сообщений. Не понимаю как передать Ref для автофокусировки
