+++ react-gb/src/components/ChatOne.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import {MessageList} from './MessageList';
-import {MessageForm} from './MessageForm';
 import Container from "@mui/material/Container";
 import {useSelector} from "react-redux";
 import {useParams} from "react-router-dom";
 import {getChat} from "../store/chat/chatSelector";
 import shallowEqual from "react-redux/lib/utils/shallowEqual";
+import {MessageFormContainer} from "./containers/MessageFormContainer";
 
 
 export function ChatOne() {
@@ -23,7 +23,7 @@ export function ChatOne() {
             <h2>Chat with: {chat.name}</h2>
             <div className="Chats">
                 <Container maxWidth={"md"}>
-                    <MessageForm/>
+                    <MessageFormContainer/>
                     <MessageList/>
                 </Container>
             </div>
